Serialization and restart of controller and table
