+++ ui/app/services/watch-list.js
@@ -29,6 +29,6 @@ export default class WatchListService extends Service {
 
   setIndexFor(url, value) {
     list[url] = +value;
-    console.log('xxx total list is now', list);
+    console.log('total list is now', list);
   }
 }
